+++ utils/http-error-handler.js
@@ -2,6 +2,7 @@ const httpError = require('./http-errors')
 const ValidationError = require('./converter/validation-error')
 const ForbiddenError = require('./converter/forbidden-error')
 const EmptyResultError = require('./converter/empty-result-error')
+const UnauthorizedError = require('./converter/unauthorized-error')
 
 function httpErrorHandler (req, res, fallbackMessage) {
   return (err) => {
@@ -19,4 +20,33 @@ function httpErrorHandler (req, res, fallbackMessage) {
   }
 }
 
-module.exports = { httpErrorHandler }
+function rpErrorMatcher (err) {
+  try {
+    const statusCode = err.response.statusCode
+    let message
+    try {
+      message = err.response.body.message
+    } catch (mesErr) {
+      message = err.message
+    }
+    switch (statusCode) {
+      case 400:
+        return new ValidationError(message)
+      case 401:
+        return new UnauthorizedError(message)
+      case 403:
+        return new ForbiddenError(message)
+      case 404:
+        return new EmptyResultError(message)
+      default:
+        return err
+    }
+  } catch (e) {
+    return err
+  }
+}
+
+module.exports = {
+  httpErrorHandler,
+  rpErrorMatcher
+}